+++ src/overtone/studio/inst.clj
@@ -14,24 +14,26 @@
         [overtone.studio.mixer]
         [overtone.studio.fx]
         [overtone.helpers.lib]
-        [overtone.libs.event])
+        [overtone.libs.event]
+        [potemkin :refer [def-map-type]])
   (:require [clojure.pprint]
+            [lentes.core :as l]
             [overtone.sc.protocols :as protocols]
             [overtone.sc.util]
-            [overtone.sc.machinery.server.comms :refer [with-server-sync]]))
+            [overtone.helpers.lib :as ov.lib]
+            [overtone.sc.machinery.server.comms :refer [with-server-sync]]
+            [overtone.libs.deps :as ov.deps]))
 
-(defonce ^{:private true} __RECORDS__
-  (do
-    (defrecord-ifn Inst [name full-name params args sdef
-                         group instance-group fx-group
-                         mixer bus fx-chain
-                         volume pan
-                         n-chans]
-      (fn [this & args]
-        (apply synth-player sdef params this [:tail instance-group] args))
+(ov.lib/defrecord-ifn-2 Inst [name full-name params args sdef
+                              ^:deref group ^:deref instance-group
+                              ^:deref fx-group
+                              ^:deref mixer mixer-params volume pan bus
+                              n-chans]
+  (fn [this & args]
+    (apply synth-player sdef params this [:tail @instance-group] args))
 
-      to-sc-id*
-      (to-sc-id [_] (to-sc-id instance-group)))))
+  to-sc-id*
+  (to-sc-id [_] (to-sc-id @instance-group)))
 
 (derive Inst :overtone.sc.node/node)
 
@@ -56,20 +58,52 @@
        out-bus 0
        volume  DEFAULT-VOLUME
        pan     DEFAULT-PAN]
-      (let [snd  (in in-bus 2)
-            snd (select (check-bad-values snd 0 0)
-                        [snd (dc 0) (dc 0) snd])
-            snd (limiter snd 0.99 0.001)
-            sndl (select 0 snd)
-            sndr (select 1 snd)]
+      (let [zero (dc 0)
+            sndl (in in-bus 1)
+            sndl (select (check-bad-values sndl 0 0)
+                         [sndl zero zero sndl])
+            sndl (limiter sndl 0.99 0.001)
+            sndr (in (+ in-bus 1) 1)
+            sndr (select (check-bad-values sndr 0 0)
+                         [sndr zero zero sndr])
+            sndr (limiter sndr 0.99 0.001)]
         (out out-bus (balance2 sndl sndr pan volume))))))
 
-(defn inst-mixer
+(defn default-get-inst-mixer
   "Instantiate a mono or stereo inst-mixer synth."
-  [n-chans & args]
+  [{:keys [n-chans] :as _inst}]
   (if (> n-chans 1)
-    (apply stereo-inst-mixer args)
-    (apply mono-inst-mixer args)))
+    stereo-inst-mixer
+    mono-inst-mixer))
+
+(defn inst-mixer
+  "Instantiate an instrument mixer."
+  [inst & args]
+  (let [get-mixer (get @studio* ::get-inst-mixer default-get-inst-mixer)
+        mixer-synth (get-mixer inst)]
+    (apply mixer-synth args)))
+
+(defn replace-inst-mixer!
+  "Replace the mixer synth in an instrument."
+  [{:keys [bus mixer-params] :as inst} new-get-inst-mixer & {:as params}]
+  (ensure-node-active! inst)
+  (let [{:keys [mixer]} (ov.lib/ov-raw-map inst)
+        mixer-synth (new-get-inst-mixer inst)
+        synth-params (->> (merge @mixer-params params)
+                          ;; Flatten to :key1 val1 :key2 val2 ...
+                          (mapcat identity))
+        new-mixer (apply mixer-synth
+                         [:replace @mixer]
+                         :in-bus bus
+                         synth-params)]
+    (reset! mixer new-mixer)))
+
+(defn replace-all-inst-mixer!
+  "Replace the mixer synth in all current and future instruments."
+  [new-get-inst-mixer & params]
+  (swap! studio* assoc ::get-inst-mixer new-get-inst-mixer)
+  (doseq [[_name inst] (:instruments @studio*)]
+    (apply replace-inst-mixer! inst new-get-inst-mixer params)))
 
 (defn inst-channels
   "Internal fn used for multimethod dispatch on Insts."
@@ -90,6 +124,13 @@
   (ensure-node-active! inst)
   (ctl (:mixer inst) :pan pan)
   (reset! (:pan inst) pan))
+
+(defn inst-mixer-ctl!
+  "Control a named parameters of the output mixer of a single instrument."
+  [inst & {:as args}]
+  (ensure-node-active! inst)
+  (apply ctl (:mixer inst) (-> args seq flatten))
+  (swap! (:mixer-params inst) merge args))
 
 (defmulti inst-fx!
   "Append an effect to an instrument channel. Returns a SynthNode or a
@@ -155,50 +196,88 @@
   [o]
   (= overtone.studio.inst.Inst (type o)))
 
+(defmacro with-server-sync-atom
+  [action-fn error-msg]
+  `(atom
+    (when (server-connected?)
+      (with-server-sync ~action-fn ~error-msg))))
+
 (defmacro inst
   [sname & args]
-  (ensure-connected!)
   `(let [[sname# full-name# params# ugens# constants# n-chans# inst-bus#] (pre-inst ~sname ~@args)
-         new-inst# (get (:instruments @studio*) full-name#)
+         new-inst# (ov.lib/ov-raw-map (get (:instruments @studio*) full-name#))
+
          container-group# (or (:group new-inst#)
-                              (with-server-sync
+                              (with-server-sync-atom
                                 #(group (str "Inst " sname# " Container")
                                         :tail (:instrument-group @studio*))
                                 "whilst creating an inst container group"))
 
-         instance-group#  (or (:instance-group new-inst#)
-                              (with-server-sync
-                                #(group (str "Inst " sname#)
-                                        :head container-group#)
-                                "whilst creating an inst instance group"))
+         instance-group# (or (:instance-group new-inst#)
+                             (with-server-sync-atom
+                               #(group (str "Inst " sname#)
+                                       :head @container-group#)
+                               "whilst creating an inst instance group"))
 
          fx-group#        (or (:fx-group new-inst#)
-                              (with-server-sync
+                              (with-server-sync-atom
                                 #(group (str "Inst " sname# " FX")
-                                        :tail container-group#)
+                                        :tail @container-group#)
                                 "whilst creating an inst fx group"))
 
          imixer#    (or (:mixer new-inst#)
-                        (inst-mixer n-chans#
-                                    [:tail container-group#]
-                                    :in-bus inst-bus#))
+                        (atom ::uninitialized-mixer))
          sdef#      (synthdef sname# params# ugens# constants#)
          arg-names# (map :name params#)
          params-with-vals# (map #(assoc % :value (control-proxy-value-atom full-name# %)) params#)
-         fx-chain#  []
-         volume#    (atom DEFAULT-VOLUME)
-         pan#       (atom DEFAULT-PAN)
+         mixer-params# (atom {:volume DEFAULT-VOLUME
+                              :pan    DEFAULT-PAN})
+         volume#    (l/derive (l/key :volume) mixer-params#)
+         pan#       (l/derive (l/key :pan) mixer-params#)
          inst#      (with-meta
                       (->Inst sname# full-name# params-with-vals# arg-names# sdef#
                               container-group# instance-group# fx-group#
-                              imixer# inst-bus# fx-chain#
-                              volume# pan#
+                              imixer# mixer-params# volume# pan# inst-bus#
                               n-chans#)
                       {:overtone.helpers.lib/to-string #(str (name (:type %)) ":" (:name %))})]
+     (when (= ::uninitialized-mixer @imixer#)
+       ;; Briefly delayed mixer creation to support passing the realized inst# to
+       ;; the mixer selection function.
+       (reset! imixer#
+               (when (server-connected?)
+                 (with-server-sync
+                   #(inst-mixer inst#
+                                [:tail @container-group#]
+                                :in-bus inst-bus#)
+                   "whilst creating an inst imixer"))))
      (load-synthdef sdef#)
      (add-instrument inst#)
      (event :new-inst :inst inst#)
      inst#))
+
+(defn- load-all-insts
+  []
+  ;; TODO Get atoms using something else.
+  (doseq [-inst (:instruments @studio*)]
+    (let [sname (first -inst)
+          {:keys [instance-group fx-group mixer bus] container-group :group}
+          (ov.lib/ov-raw-map (second -inst))]
+      (reset! container-group (group (str "Inst " sname " Container")
+                                     :tail (:instrument-group @studio*)))
+
+      (reset! instance-group (group (str "Inst " sname)
+                                    :head @container-group))
+
+      (reset! fx-group (group (str "Inst " sname " FX")
+                              :tail @container-group))
+
+      (reset! mixer (inst-mixer (second -inst)
+                                [:tail @container-group]
+                                :in-bus bus))))
+
+  (ov.deps/satisfy-deps :insts-loaded))
+
+(ov.deps/on-deps :studio-setup-completed ::load-all-insts load-all-insts)
 
 (defmacro definst
   "Define an instrument and return a player function. The instrument
@@ -275,8 +354,8 @@
     (inst-pan! bar 1 -1)   ;ch1 right, ch2 left.
     (inst-volume! bar 0 1) ;mute ch1.
 
-  * Each instrument has an fx-chain to which you can add any number of
-    'fx synths' using the inst-fx function.
+  * Each instrument has an fx-group to which you can add any number of
+    'fx synths' using the inst-fx! function.
   "
   {:arglists '([name doc-string? params ugen-form])}
   [i-name & inst-form]
@@ -324,6 +403,7 @@
                               (with-out-str (doseq [n sub-nodes] (pr n))))))))
 
 (extend Inst
+
   ISynthNode
   {:node-free  node-free*
    :node-pause node-pause*